pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-pset-solver"
version = "0.1.0"
description = "A Python script to automatically solve problem sets from PDF files using LLMs."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "anthropic",
    "google-genai",
    "openai",
    "python-dotenv",
    "PyMuPDF",
    "PyMuPDF4LLM",
    "tenacity",
    "httpx",
    "requests",
    "pdfplumber",
    "Pillow"
]

[tool.setuptools.packages.find]
include = ["llm_providers*", "utils*"]

[project.scripts]
auto-pset = "main:main"

[project.urls]
"Homepage" = "https://github.com/example/auto-pset"
"Bug Tracker" = "https://github.com/example/auto-pset/issues" 